fix: handle multiple params lists in for infer type
